backend/requirements.txt
fastapi
uvicorn
pillow
tweepy
python-dotenv
python-multipart
itsdangerous
